[frontend] Rework navbar & logout button for client [SSR] partials for views [CORS] cfg for CORS origin

+++ backend/src/api/models/sale.model.js
@@ -14,7 +14,7 @@ const Sale = sequelize.define('sale', {
         allowNull : false
     },
     total : {
-        type : DataTypes.INTEGER,
+        type : DataTypes.DOUBLE(15,2),
         allowNull : false
     }
 }, {tableName : 'sales', timestamps : true, updatedAt: false})
